Correções tela atividades
