+++ src/components/atividade-card/AtividadeCard.jsx
@@ -21,7 +21,7 @@ const AtividadeCard = ({ atividade, trigger_reload }) => {
                 </div>
                 <div className='flex flex-col md:w-[20vw] pb-4'>
                     <p className="truncate">{atividade.enunciado}</p>
-                    <p>{atividade.turma == undefined ? 'Sem Turma' : atividade.turma.nome}</p>
+                    <p className="truncate">{atividade.turma == undefined ? 'Sem Turma' : atividade.turma.nome}</p>
                     {/* <p className="truncate opacity-70">{atividade}</p> */}
                 </div>
                 <div className='flex flex-col md:w-[10vw] pb-4'>
@@ -33,7 +33,7 @@ const AtividadeCard = ({ atividade, trigger_reload }) => {
                         >
                             <path d="M18 10H9V7c0-1.654 1.346-3 3-3s3 1.346 3 3h2c0-2.757-2.243-5-5-5S7 4.243 7 7v3H6a2 2 0 00-2 2v8a2 2 0 002 2h12a2 2 0 002-2v-8a2 2 0 00-2-2zm-7.939 5.499A2.002 2.002 0 0114 16a1.99 1.99 0 01-1 1.723V20h-2v-2.277a1.992 1.992 0 01-.939-2.224z" />
                         </svg>
-                        <p>{formatDateTime(atividade.data_abertura)}</p>
+                        <p className="truncate">{formatDateTime(atividade.data_abertura)}</p>
                         {/* <p className="truncate">{formatDateTime(atividade.data_abertura)}</p> */}
                     </div>
                     <div className='flex flex-row gap-2 mt-1'>
@@ -44,7 +44,7 @@ const AtividadeCard = ({ atividade, trigger_reload }) => {
                         >
                         <path d="M12 2C9.243 2 7 4.243 7 7v3H6a2 2 0 00-2 2v8a2 2 0 002 2h12a2 2 0 002-2v-8a2 2 0 00-2-2h-1V7c0-2.757-2.243-5-5-5zM9 7c0-1.654 1.346-3 3-3s3 1.346 3 3v3H9V7zm4 10.723V20h-2v-2.277a1.993 1.993 0 01.567-3.677A2.001 2.001 0 0114 16a1.99 1.99 0 01-1 1.723z" />
                         </svg>
-                        <p>{formatDateTime(atividade.data_encerramento)}</p>
+                        <p className="truncate">{formatDateTime(atividade.data_encerramento)}</p>
                     </div>
                 </div>
                 <div className='flex md:flex-row flex-col md:w-[20vw] justify-between'>
