+++ src/components/atividade-card/AtividadeVerModalAdmin.jsx
@@ -1,27 +1,27 @@
 import axios from 'axios'
 import React, { useState } from 'react'
 import { API_BASE_URL } from '../../util/constants'
-import { ToastifyNotificate } from '../toast/Toastify'
-import { ToastContainer } from 'react-toastify'
 
 const AtividadeVerModalAdmin = ({ atividade, trigger_reload }) => {
     const [ nome, setNome ] = useState(atividade.nome)
     const [ enunciado, setEnunciado ] = useState(atividade.enunciado)
-    const data_abertura_formatted = new Date(atividade.data_abertura).toLocaleDateString()
-    const data_encerramento_formatted = new Date(atividade.data_encerramento).toLocaleDateString()
+    const [ data_abertura, setDataAbertura ] = useState(atividade.data_abertura)
+    const [ data_encerramento, setDataEncerramento ] = useState(atividade.data_encerramento)
+
     const update_atividade = async() => {
         await axios.patch(`${API_BASE_URL}/atividade/update`, {
             id: atividade._id,
             nome: nome,
-            enunciado: enunciado
+            enunciado: enunciado,
+            data_abertura: data_abertura,
+            data_encerramento: data_encerramento
         }).then((res) => {
-            ToastifyNotificate({ message: res.data.message, type: 'success' })
             trigger_reload();
             document.getElementById(atividade._id).close()
         }).catch((err) => {
-            ToastifyNotificate({ message: err.response.data.message, type: 'error' })
         })
     }
+
     return (
         <dialog id={atividade._id} className="modal">
             <div className="modal-box flex flex-col md:gap-y-4 gap-y-3 max-w-screen-lg max-h-[88vh] bg-base-200">
@@ -34,22 +34,36 @@ const AtividadeVerModalAdmin = ({ atividade, trigger_reload }) => {
                 
                 <div className='flex flex-col w-full'>
                     <p className="text-lg font-extrabold opacity-70 pb-0.5 pt-1">Enunciado:</p>
-                    <textarea onChange={e => setEnunciado(e.target.value)} style={{ resize: 'none' }} className="w-full h-36 p-2 bg-base-300 rounded-lg" value={atividade.enunciado}></textarea>
+                    <textarea onChange={e => setEnunciado(e.target.value)} style={{ resize: 'none' }} className="w-full h-36 p-2 bg-base-300 rounded-lg" value={enunciado}></textarea>
                 </div>
 
-                <div className='flex flex-col w-full'>
-                    <p className="text-lg font-extrabold opacity-70 pb-0.5 pt-1">Data de abertura:</p>
-                    <p className="text-lg font-extrabold">{data_abertura_formatted}</p>
-                    <p className="text-lg font-extrabold opacity-70 pb-0.5 pt-1">Data de encerramento:</p>
-                    <p className="text-lg font-extrabold">{data_encerramento_formatted}</p>
-                </div>
+                <label className='form-control'>
+                    <div className='label'>
+                        <span className='label-text'>Data de Abertura:</span>
+                    </div>
+                    <input 
+                        type='datetime-local' 
+                        className='input input-bordered' 
+                        onChange={e => setDataAbertura(e.target.value)}
+                    />
+                </label>
+
+                <label className='form-control'>
+                    <div className='label'>
+                        <span className='label-text'>Data de Encerramento:</span>
+                    </div>
+                    <input 
+                        type='datetime-local' 
+                        className='input input-bordered' 
+                        onChange={e => setDataEncerramento(e.target.value)}
+                    />
+                </label>
 
                 <button onClick={() => update_atividade()} className='btn btn-primary text-white'>Editar</button>
             </div>
             <form method="dialog" className="modal-backdrop">
                 <button></button>
             </form>
-            <ToastContainer />
         </dialog>
     )
 }
